Remove data manipulation from request migration
The migration for making request a first class objects has been run on
all deployed instances and thus is no longer needed. The issue is that
it makes changing the Request model impossible and we need to remove
this part in order to make modifications to Requests.

Since this migration should from now on only be run on an empty
database, It should not make any difference if the removed code exists
or not.